# src/main/java/com/afkl/travel/exercise/model/LocationType.java
package com.afkl.travel.exercise.model;

public class LocationType {

	public static final String country = "COUNTRY";
    public static final String city="CITY";
	public static final String airport= "AIRPORT";
}